backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
drf-spectacular==0.27.0
requests==2.31.0
django-filter==23.5
gunicorn==21.2.0
whitenoise==6.6.0
setuptools
groq
